Move classes into monsterrace package
