+++ README.md
@@ -75,3 +75,4 @@
 15. rankMap을 생성자에서 생성하도록 코드를 수정하였습니다.
 16. Race 클래스의 책임이 분명해지도록 클래스명을 리팩토링 하였습니다.
 17. Game 클래스의 인스턴스 변수를 private로 접근 제어자를 설정해주었습니다. (왜 뺴먹었지...)
+18. 소스가 있는 패키지의 위치를 이동하였습니다.
