+++ src/main/java/dev/idion/monsterrace/Monster.java
@@ -1,4 +1,4 @@
-package dev.idion;
+package dev.idion.monsterrace;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
