+++ src/main/java/dev/idion/monsterrace/StringConstants.java
@@ -1,4 +1,4 @@
-package dev.idion;
+package dev.idion.monsterrace;
 
 public enum StringConstants {
     PROMPT("> "),
